cli: update deploy

+++ packages/example/src/types/ERC20.ts
@@ -3,7 +3,7 @@ export declare class ERC20 {
   static load(address: Address, chainId: u64): ERC20;
   name(): string;
   totalSupply(): BigInt;
-  decimals(): BigInt;
+  decimals(): u8;
   balanceOf(_owner: Address): BigInt;
   symbol(): string;
   allowance(_owner: Address, _spender: Address): BigInt;
